+++ engine/flink/api/src/main/scala/pl/touk/nussknacker/engine/flink/api/typeinformation/TypeInformationDetection.scala
@@ -19,6 +19,6 @@ trait TypeInformationDetection extends Serializable {
 
   def forValueWithContext[T](validationContext: ValidationContext, value: TypingResult): TypeInformation[ValueWithContext[T]]
 
-  def forType(typingResult: TypingResult): TypeInformation[Any]
+  def forType(typingResult: TypingResult): TypeInformation[AnyRef]
 
 }